webapp/requirements.txt
pytest==7.4.3
mysql-connector-python==8.2.0
flask==3.0.0
